I have completed the UserDefaults actions of retrieving, adding, saving and deleting favourites.
This allows the user to have a list separate from the search results. The user can easily navigate to the list via the tab bar.

It is handled by a Persistence Manager and a Favourites Coordinator.

The “Add to favourites” button title label in MovieDetailVC changes when tapped to tell the user the movie was added to the favourites list. The delete action is shown via swipe to delete on the table view cell in FavouritesListVC.

To be added - an alert to tell the user if the UserDefaults actions fail.